+++ lindera/src/analyzer.rs
@@ -11,6 +11,9 @@ use crate::{
     },
     error::LinderaErrorKind,
     token_filter::{
+        japanese_katakana_stem::{
+            JapaneseKatakanaStemTokenFilter, JAPANESE_KATAKANA_STEM_TOKEN_FILTER_NAME,
+        },
         length::{LengthTokenFilter, LENGTH_TOKEN_FILTER_NAME},
         stop_words::{StopWordsTokenFilter, STOP_WORDS_TOKEN_FILTER_NAME},
     },
@@ -118,6 +121,11 @@ impl Analyzer {
                             token_filters
                                 .push(Box::new(LengthTokenFilter::from_slice(&args_bytes)?));
                         }
+                        JAPANESE_KATAKANA_STEM_TOKEN_FILTER_NAME => {
+                            token_filters.push(Box::new(
+                                JapaneseKatakanaStemTokenFilter::from_slice(&args_bytes)?,
+                            ));
+                        }
                         _ => {
                             return Err(LinderaErrorKind::Deserialize.with_error(anyhow::anyhow!(
                                 "unknown token filter {}.",
@@ -137,7 +145,7 @@ impl Analyzer {
             character_filter.apply(text)?;
         }
 
-        let mut tokens = self.tokenizer.tokenize(text.as_str())?;
+        let mut tokens = self.tokenizer.tokenize_with_details(text.as_str())?;
 
         for token_filter in &self.token_filters {
             token_filter.apply(&mut tokens)?;
@@ -167,19 +175,15 @@ mod tests {
                     "kind": "mapping",
                     "args": {
                         "mapping": {
-                            "ｱ": "ア",
-                            "ｲ": "イ",
-                            "ｳ": "ウ",
-                            "ｴ": "エ",
-                            "ｵ": "オ"
+                            "(株)": "株式会社"
                         }            
                     }
                 },
                 {
                     "kind": "regex",
                     "args": {
-                        "pattern": "リンデラ",
-                        "replacement": "lindera"
+                        "pattern": "\\s{2,}",
+                        "replacement": " "
                     }
                 }
             ],
@@ -187,61 +191,33 @@ mod tests {
                 "dictionary": {
                     "kind": "ipadic"
                 },
-                "mode": {
-                    "decompose": {
-                        "kanji_penalty_length_threshold": 2,
-                        "kanji_penalty_length_penalty": 3000,
-                        "other_penalty_length_threshold": 7,
-                        "other_penalty_length_penalty": 1700
-                    }
-                }
+                "mode": "normal"
             },
             "token_filters": [
                 {
                     "kind": "stop_words",
                     "args": {
                         "stop_words": [
-                            "a",
-                            "an",
-                            "and",
-                            "are",
-                            "as",
-                            "at",
                             "be",
-                            "but",
-                            "by",
-                            "for",
-                            "if",
-                            "in",
-                            "into",
                             "is",
-                            "it",
-                            "no",
                             "not",
-                            "of",
-                            "on",
                             "or",
-                            "such",
-                            "that",
                             "the",
-                            "their",
-                            "then",
-                            "there",
-                            "these",
-                            "they",
                             "this",
-                            "to",
-                            "was",
-                            "will",
-                            "with"
+                            "to"
                         ]
                     }
                 },
                 {
                     "kind": "length",
                     "args": {
-                        "min": 1,
-                        "max": 3
+                        "min": 1
+                    }
+                },
+                {
+                    "kind": "japanese_katakana_stem",
+                    "args": {
+                        "min": 3
                     }
                 }
             ]
@@ -265,10 +241,18 @@ mod tests {
                     }
                 },
                 {
+                    "kind": "mapping",
+                    "args": {
+                        "mapping": {
+                            "(株)": "株式会社"
+                        }            
+                    }
+                },
+                {
                     "kind": "regex",
                     "args": {
-                        "pattern": "リンデラ",
-                        "replacement": "lindera"
+                        "pattern": "\\s{2,}",
+                        "replacement": " "
                     }
                 }
             ],
@@ -276,20 +260,33 @@ mod tests {
                 "dictionary": {
                     "kind": "ipadic"
                 },
-                "mode": {
-                    "decompose": {
-                        "kanji_penalty_length_threshold": 2,
-                        "kanji_penalty_length_penalty": 3000,
-                        "other_penalty_length_threshold": 7,
-                        "other_penalty_length_penalty": 1700
-                    }
-                }
+                "mode": "normal"
             },
             "token_filters": [
+                {
+                    "kind": "stop_words",
+                    "args": {
+                        "stop_words": [
+                            "be",
+                            "is",
+                            "not",
+                            "or",
+                            "the",
+                            "this",
+                            "to"
+                        ]
+                    }
+                },
                 {
                     "kind": "length",
                     "args": {
                         "min": 2
+                    }
+                },
+                {
+                    "kind": "japanese_katakana_stem",
+                    "args": {
+                        "min": 3
                     }
                 }
             ]
@@ -297,12 +294,12 @@ mod tests {
         "#;
         let analyzer = Analyzer::from_slice(config_str.as_bytes()).unwrap();
 
-        let mut text = "ﾘﾝﾃﾞﾗは、日本語の形態素解析エンジンです。".to_string();
+        let mut text = "Ｌｉｎｄｅｒａは、日本語の形態素解析ｴﾝｼﾞﾝです。".to_string();
         let tokens = analyzer.analyze(&mut text).unwrap();
 
         assert_eq!(
             tokens.iter().map(|t| t.text).collect::<Vec<_>>(),
-            vec!["lindera", "日本語", "形態素", "解析", "エンジン", "です"]
+            vec!["Lindera", "日本語", "形態素", "解析", "エンジン", "です"]
         );
     }
 
@@ -322,16 +319,15 @@ mod tests {
                     "kind": "mapping",
                     "args": {
                         "mapping": {
-                            "（株）": "株式会社",
-                            "〒": "郵便"
+                            "(株)": "株式会社"
                         }            
                     }
                 },
                 {
                     "kind": "regex",
                     "args": {
-                        "pattern": "リンデラ",
-                        "replacement": "lindera"
+                        "pattern": "\\s{2,}",
+                        "replacement": " "
                     }
                 }
             ],
@@ -339,35 +335,33 @@ mod tests {
                 "dictionary": {
                     "kind": "ipadic"
                 },
-                "mode": {
-                    "decompose": {
-                        "kanji_penalty_length_threshold": 2,
-                        "kanji_penalty_length_penalty": 3000,
-                        "other_penalty_length_threshold": 7,
-                        "other_penalty_length_penalty": 1700
-                    }
-                }
+                "mode": "normal"
             },
             "token_filters": [
                 {
                     "kind": "stop_words",
                     "args": {
                         "stop_words": [
-                            "a",
-                            "an",
-                            "and",
-                            "are",
-                            "as",
-                            "at",
-                            "with"
+                            "be",
+                            "is",
+                            "not",
+                            "or",
+                            "the",
+                            "this",
+                            "to"
                         ]
                     }
                 },
                 {
                     "kind": "length",
                     "args": {
-                        "min": 1,
-                        "max": 3,  // wrong
+                        "min": 2
+                    }
+                },
+                {
+                    "kind": "japanese_katakana_stem_wrong",  // wrong token filter name
+                    "args": {
+                        "min": 3
                     }
                 }
             ]