+++ lindera/src/token_filter.rs
@@ -1,2 +1,3 @@
+pub mod japanese_katakana_stem;
 pub mod length;
 pub mod stop_words;